services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0
python-dotenv==1.0.0
pgvector==0.2.4
sentence-transformers==2.2.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
ocrmypdf==16.0.4
pypdf==3.17.4
pdfminer.six==20231228
unstructured[pdf]==0.11.8
pillow==10.2.0
numpy==1.26.3
ollama==0.1.7
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
mypy==1.8.0